Restructured view classes and made MainFrame shorter. Moved handlers to controller.
