+++ src/view/creators/PortalCreator.java
@@ -1,8 +1,9 @@
-package view;
+package view.creators;
 
 import controller.GameLogic;
 import controller.ImageResizer;
-import view.handlers.ActionHandler;
+import view.MainFrame;
+import controller.handlers.ActionHandler;
 
 import javax.swing.*;
 import java.awt.event.MouseEvent;
