+++ src/view/panels/FinalScenePanel.java
@@ -1,7 +1,8 @@
-package view;
+package view.panels;
 
 import controller.ImageResizer;
-import view.handlers.ActionHandler;
+import view.MainFrame;
+import controller.handlers.ActionHandler;
 
 import javax.swing.*;
 import java.awt.*;
